% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runsost.R
\name{runsost}
\alias{runsost}
\title{perform structure identification on a selected image}
\usage{
runsost(
  spe,
  imageId = "E03",
  catname = "cell_category",
  markSel = "islet",
  imageColName = "image_name",
  dim.recon = 500,
  lwd = 2
)
}
\arguments{
\item{spe}{SpatialExperiment instance}

\item{imageId}{character(1) used for filtering}

\item{catname}{character(1) used to label cells}

\item{markSel}{character(1) a colData component identifying structure}

\item{imageColName}{character(1) name of colData component used for filtering}

\item{dim.recon}{numeric(1) passed to sosta::reconstructShapeDensityImage}

\item{lwd}{used for plotting boundary of structure}
}
\description{
perform structure identification on a selected image
}
