% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyDamond.R
\name{surveyDamond}
\alias{surveyDamond}
\title{shiny app for running sosta on any image in the Damond SPE from sosta package}
\usage{
surveyDamond(
  spe,
  qnassayName = "quant_norm",
  catname = "cell_category",
  markSel = "islet",
  imageColName = "image_name",
  dim.recon = 500,
  lwd = 2
)
}
\arguments{
\item{spe}{SpatialExperiment instance}

\item{catname}{character(1) used to label cells}

\item{markSel}{character(1) a colData component identifying structure}

\item{imageColName}{character(1) name of colData component used for filtering}

\item{dim.recon}{numeric(1) passed to sosta::reconstructShapeDensityImage}

\item{lwd}{used for plotting boundary of structure}

\item{qnasssayName}{character(1) defaults to "quant_norm"}
}
\description{
shiny app for running sosta on any image in the Damond SPE from sosta package
}
\examples{
# NOTE: run example(surveyDamond, ask=FALSE) !!
# NOTE: select target 'INS' for good view of structures identified with
# default settings
if (interactive()) {
 requireNamespace("AnnotationHub")
 requireNamespace("ExperimentHub")
 eh <- try(ExperimentHub::ExperimentHub())
 if (inherits(eh, "try-error")) 
     eh = ExperimentHub::ExperimentHub(localHub=TRUE)
 qout = AnnotationHub::query(eh, "Damond_2019_Pancreas - sce - v1 - full")
 stopifnot(length(qout)==1)
 oid = names(qout)
 # Load single cell experiment object
 spe <- eh[[oid]]
 # Convert to spatial experiment object
 spe <- toSpatialExperiment(spe,
     sample_id = "image_name",
     spatialCoordsNames = c("cell_x", "cell_y"))
 surveyDamond(spe)
}
}
